Fix : warnings. convert to unsigned int

+++ LYCategory/_Foundation/_Crypto/NSString+Encrypt.m
@@ -49,7 +49,7 @@
 	
     unsigned char digest[16];
 	
-    CC_MD5(cstr, strlen(cstr), digest);
+    CC_MD5(cstr, (unsigned int)strlen(cstr), digest);
 	
     return [NSString stringWithFormat:
 			@"%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X",
@@ -67,7 +67,7 @@
 	
 	uint8_t digest[CC_SHA1_DIGEST_LENGTH];
 	
-	CC_SHA1(data.bytes, data.length, digest);
+	CC_SHA1(data.bytes, (unsigned int)data.length, digest);
 	
 	NSMutableString* output = [NSMutableString stringWithCapacity:CC_SHA1_DIGEST_LENGTH * 2];
